+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 
 
-- **Important**: It is recommended to run `occ memories:places-setup` again if you use MySQL or MariaDB
+- **Important**: You must run `occ memories:places-setup` again after this update
 - Hide files starting with `.` in the timeline
 - Prevent automatically retrying failed indexing jobs
 - Support for 3GP videos ([#1055](https://github.com/pulsejet/memories/issues/1055))